requirements-dev.txt
-r requirements.txt

codecov==1.6.3
django-autofixture==0.12.0
django-debug-toolbar==1.4
flake8-blind-except==0.1.0
flake8-deprecated==1.0
flake8-import-order==0.7
flake8-pep257==1.0.5
flake8-pep3101==0.3
flake8-print==2.0.2
flake8-quotes==0.2.4
flake8-strict==0.1.2
flake8-tidy-imports==1.0.0
pep8-naming==0.3.3
